Intro / FAQs
DINO aims to become the go-to decentralised platform by providing an array of services and features that will enable users of all background to participate in the evolution of decentralised finance. The DINO platform is intuitive, robust and user-friendly as it was developed with security and the community in mind. As a one-stop platform, DINO aims to attract existing as well as new users to blockchain and De-Fi by integrating DEX liquidity, offering Staking Support and other innovative features that make the DINO platform the preferential De-Fi application on HECO and other chains.
 
Use-cases of DINO are varied and form the DINO Ecosystem which will be fully operational once all features are released on the platform. DINO visualises a thriving ecosystem whereby the $DINO token is a clear representation of the value of the DINO platform and this will be achieved in part by strategic co-operations and shared visions with other platforms and service-providers.